// src/nps/nps.enums.ts
export enum NPSTier {
  TIERI = 'Tier I',
  TIERII = 'Tier II',
}

export enum FundManager {
  HDFC = 'HDFC Pension Management Company Limited',
  ABSL = 'Aditya Birla Sun Life Pension Management Limited',
  UTI = 'UTI Retirement Solutions Limited',
  SBI = 'SBI Pension Funds Private Limited',
  ICICI = 'ICICI Prudential Pension Funds Management Company Limited',
  KOTAK = 'Kotak Mahindra Pension Fund Limited',
  LIC = 'LIC Pension Fund',
}

export enum AssetClass {
  E = 'Equity (E)',
  C = 'Corporate Bonds (C)',
  G = 'Government Securities (G)',
  A = 'Alternative Assets (A)'
}

export enum TransactionType {
  CR = 'Credit',
  DR = 'Debit',
}